refactor getting ready
