+++ src/java/org/apache/cassandra/service/accord/serializers/TopologySerializers.java
@@ -20,10 +20,6 @@ package org.apache.cassandra.service.accord.serializers;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.util.ArrayList;
-import java.util.BitSet;
-import java.util.List;
-import java.util.Objects;
 
 import accord.local.Node;
 import accord.primitives.Range;
@@ -38,11 +34,8 @@ import org.apache.cassandra.io.util.DataInputPlus;
 import org.apache.cassandra.io.util.DataOutputPlus;
 import org.apache.cassandra.schema.TableId;
 import org.apache.cassandra.service.accord.TokenRange;
-import org.apache.cassandra.service.accord.api.TokenKey;
 import org.apache.cassandra.utils.ArraySerializers;
-import org.apache.cassandra.utils.BitSetSerializer;
 import org.apache.cassandra.utils.CollectionSerializers;
-import org.apache.cassandra.utils.ImmutableUniqueList;
 
 public class TopologySerializers
 {
@@ -139,310 +132,6 @@ public class TopologySerializers
             return size;
         }
     }
-
-    public static class CompactTopology
-    {
-        private static class Range
-        {
-            final TokenKey start;
-            final TokenKey end;
-
-            private Range(TokenKey start, TokenKey end)
-            {
-                this.start = start.withTable(EMPTY);
-                this.end = end.withTable(EMPTY);
-            }
-
-            private Range(TokenRange other)
-            {
-                this(other.start(), other.end());
-            }
-
-            @Override
-            public boolean equals(Object o)
-            {
-                if (o == null || getClass() != o.getClass()) return false;
-                Range range = (Range) o;
-                return Objects.equals(start, range.start) && Objects.equals(end, range.end);
-            }
-
-            @Override
-            public int hashCode()
-            {
-                return Objects.hash(start, end);
-            }
-        }
-
-        private static final class RangeSerializer implements UnversionedSerializer<CompactTopology.Range>
-        {
-            private static final RangeSerializer instance = new RangeSerializer();
-
-            @Override
-            public void serialize(CompactTopology.Range t, DataOutputPlus out) throws IOException
-            {
-                TokenKey.noTableSerializer.serialize(t.start, out);
-                TokenKey.noTableSerializer.serialize(t.end, out);
-            }
-
-            @Override
-            public CompactTopology.Range deserialize(DataInputPlus in) throws IOException
-            {
-                return new CompactTopology.Range(TokenKey.noTableSerializer.deserialize(EMPTY, in),
-                                                 TokenKey.noTableSerializer.deserialize(EMPTY, in));
-            }
-
-            @Override
-            public long serializedSize(CompactTopology.Range t)
-            {
-                return TokenKey.noTableSerializer.serializedSize(t.start)
-                       + TokenKey.noTableSerializer.serializedSize(t.end);
-            }
-        }
-
-        private static BitSet bitSet(SortedArrayList<Node.Id> src, SortedArrayList<Node.Id> subset)
-        {
-            BitSet bitSet = new BitSet(src.size());
-            for (int i = 0; i < src.size(); i++)
-            {
-                if (subset.contains(src.get(i)))
-                    bitSet.set(i);
-            }
-            return bitSet;
-        }
-
-        private static SortedArrayList<Node.Id> fromBitSet(SortedArrayList<Node.Id> src, BitSet bitSet)
-        {
-            SortedArrayList.Builder<Node.Id> builder = new SortedArrayList.Builder<>(new Node.Id[bitSet.cardinality()]);
-            for (int i = 0; i < src.size(); i++)
-            {
-                if (bitSet.get(i))
-                    builder.add(src.get(i));
-            }
-            return builder.build();
-        }
-
-        private static class ShardRef
-        {
-            final int tableIdx;
-            final int rangeIdx;
-            final SortedArrayList<Node.Id> nodes;
-            final BitSet notInFastPath;
-            final BitSet joining;
-            final TinyEnumSet<Shard.Flag> flags;
-
-            private ShardRef(int tableIdx, int rangeIdx,
-                             SortedArrayList<Node.Id> nodes,
-                             SortedArrayList<Node.Id> notInFastPath,
-                             SortedArrayList<Node.Id> joining,
-                             TinyEnumSet<Shard.Flag> flags)
-            {
-                this.tableIdx = tableIdx;
-                this.rangeIdx = rangeIdx;
-                this.nodes = nodes;
-                this.notInFastPath = bitSet(nodes, notInFastPath);
-                this.joining = bitSet(nodes, joining);
-                this.flags = flags;
-            }
-
-            private ShardRef(int tableIdx, int rangeIdx, SortedArrayList<Node.Id> nodes, BitSet notInFastPath, BitSet joining, TinyEnumSet<Shard.Flag> flags)
-            {
-                this.tableIdx = tableIdx;
-                this.rangeIdx = rangeIdx;
-                this.nodes = nodes;
-                this.notInFastPath = notInFastPath;
-                this.joining = joining;
-                this.flags = flags;
-            }
-
-            public Shard shard(TokenRange range)
-            {
-                return Shard.SerializerSupport.create(range, nodes, fromBitSet(nodes, notInFastPath), fromBitSet(nodes, joining), flags);
-            }
-
-            @Override
-            public boolean equals(Object o)
-            {
-                if (o == null || getClass() != o.getClass()) return false;
-                ShardRef shardRef = (ShardRef) o;
-                return tableIdx == shardRef.tableIdx && rangeIdx == shardRef.rangeIdx && Objects.equals(nodes, shardRef.nodes) && Objects.equals(notInFastPath, shardRef.notInFastPath) && Objects.equals(joining, shardRef.joining) && Objects.equals(flags, shardRef.flags);
-            }
-
-            @Override
-            public int hashCode()
-            {
-                return Objects.hash(tableIdx, rangeIdx, nodes, notInFastPath, joining, flags);
-            }
-        }
-
-        private final long epoch;
-        private final SortedArrayList<Node.Id> staleNodes;
-        private final ImmutableUniqueList<TableId> tables;
-        private final ImmutableUniqueList<Range> ranges;
-        private final List<ShardRef> shards;
-
-        CompactTopology(long epoch,
-                        SortedArrayList<Node.Id> staleNodes,
-                        ImmutableUniqueList<TableId> tables,
-                        ImmutableUniqueList<Range> ranges,
-                        List<ShardRef> shards)
-        {
-            this.epoch = epoch;
-            this.staleNodes = staleNodes;
-            this.tables = tables;
-            this.ranges = ranges;
-            this.shards = shards;
-        }
-
-        public CompactTopology(Topology topology)
-        {
-            epoch = topology.epoch();
-            staleNodes = topology.staleIds();
-            ImmutableUniqueList.Builder<TableId> tablesBuilder = ImmutableUniqueList.builder();
-            ImmutableUniqueList.Builder<Range> rangesBuilder = ImmutableUniqueList.builder();
-            List<ShardRef> shards = new ArrayList<>(topology.size());
-            for (Shard shard : topology.shards())
-            {
-                TokenRange range = (TokenRange) shard.range;
-                tablesBuilder.add(range.table());
-                Range r = new Range(range);
-                rangesBuilder.add(r);
-
-                int tableIdx = tablesBuilder.indexOf(range.table());
-                int rangeIdx = rangesBuilder.indexOf(r);
-                shards.add(new ShardRef(tableIdx, rangeIdx, shard.nodes, shard.notInFastPath, shard.joining, shard.flags()));
-            }
-            this.tables = tablesBuilder.buildAndClear();
-            this.ranges = rangesBuilder.buildAndClear();
-            this.shards = shards;
-        }
-
-        public Topology topology()
-        {
-            Shard[] shards = shards();
-            return new Topology(epoch, staleNodes, shards);
-        }
-
-        private Shard[] shards()
-        {
-            Shard[] result = new Shard[this.shards.size()];
-            for (int i = 0; i < result.length; i++)
-            {
-                ShardRef ref = this.shards.get(i);
-                TokenRange range = tokenRange(ref.tableIdx, ref.rangeIdx);
-
-                result[i] = ref.shard(range);
-            }
-            return result;
-        }
-
-        private TokenRange tokenRange(int tableIdx, int rangeIdx)
-        {
-            TableId tableId = tableId(tableIdx);
-            Range range = ranges.get(rangeIdx);
-            return TokenRange.createUnsafe(range.start.withTable(tableId), range.end.withTable(tableId));
-        }
-
-        private TableId tableId(int tableIdx)
-        {
-            return tables.get(tableIdx);
-        }
-
-        @Override
-        public boolean equals(Object o)
-        {
-            if (o == null || getClass() != o.getClass()) return false;
-            CompactTopology that = (CompactTopology) o;
-            return epoch == that.epoch && Objects.equals(staleNodes, that.staleNodes) && Objects.equals(tables, that.tables) && Objects.equals(ranges, that.ranges) && Objects.equals(shards, that.shards);
-        }
-
-        @Override
-        public int hashCode()
-        {
-            return Objects.hash(epoch, staleNodes, tables, ranges, shards);
-        }
-    }
-
-    public static final UnversionedSerializer<CompactTopology> dictionaryTopology = new UnversionedSerializer<CompactTopology>()
-    {
-        @Override
-        public void serialize(CompactTopology topology, DataOutputPlus out) throws IOException
-        {
-            out.writeLong(topology.epoch);
-            CollectionSerializers.serializeList(topology.staleNodes, out, nodeId);
-
-            CollectionSerializers.serializeList(topology.tables, out, TableId.compactComparableSerializer);
-            CollectionSerializers.serializeList(topology.ranges, out, CompactTopology.RangeSerializer.instance);
-
-            out.writeUnsignedVInt32(topology.shards.size());
-            for (CompactTopology.ShardRef shard : topology.shards)
-            {
-                out.writeUnsignedVInt32(shard.tableIdx);
-                out.writeUnsignedVInt32(shard.rangeIdx);
-
-                //TODO (perf): can this be compressed?
-                CollectionSerializers.serializeList(shard.nodes, out, nodeId);
-                BitSetSerializer.instance.serialize(shard.notInFastPath, out);
-                BitSetSerializer.instance.serialize(shard.joining, out);
-                out.writeUnsignedVInt32(shard.flags.bitset());
-            }
-        }
-
-        @Override
-        public long serializedSize(CompactTopology topology)
-        {
-            long size = Long.BYTES;
-            size += CollectionSerializers.serializedListSize(topology.staleNodes, nodeId);
-
-            size += CollectionSerializers.serializedListSize(topology.tables, TableId.compactComparableSerializer);
-            size += CollectionSerializers.serializedListSize(topology.ranges, CompactTopology.RangeSerializer.instance);
-
-            ImmutableUniqueList<TableId> tables = topology.tables;
-            ImmutableUniqueList<CompactTopology.Range> ranges = topology.ranges;
-
-            size += TypeSizes.sizeofUnsignedVInt(topology.shards.size());
-            for (CompactTopology.ShardRef shard : topology.shards)
-            {
-                size += TypeSizes.sizeofUnsignedVInt(shard.tableIdx);
-                size += TypeSizes.sizeofUnsignedVInt(shard.rangeIdx);
-
-                size += CollectionSerializers.serializedListSize(shard.nodes, nodeId);
-                size += BitSetSerializer.instance.serializedSize(shard.notInFastPath);
-                size += BitSetSerializer.instance.serializedSize(shard.joining);
-                size += TypeSizes.sizeofUnsignedVInt(shard.flags.bitset());
-            }
-            return size;
-        }
-
-        @Override
-        public CompactTopology deserialize(DataInputPlus in) throws IOException
-        {
-            long epoch = in.readLong();
-            SortedArrayList<Node.Id> staleNodes = SortedArrayList.copySorted(CollectionSerializers.deserializeList(in, nodeId), Node.Id[]::new);
-
-            ImmutableUniqueList<TableId> tables = ImmutableUniqueList.copyOf(CollectionSerializers.deserializeList(in, TableId.compactComparableSerializer));
-            ImmutableUniqueList<CompactTopology.Range> ranges = ImmutableUniqueList.copyOf(CollectionSerializers.deserializeList(in, CompactTopology.RangeSerializer.instance));
-
-            int size = in.readUnsignedVInt32();
-            List<CompactTopology.ShardRef> shards = new ArrayList<>();
-            for (int i = 0; i < size; i++)
-            {
-                int tableIndex = in.readUnsignedVInt32();
-                int rangeIndex = in.readUnsignedVInt32();
-                // confirm indexes are valid
-                //noinspection ResultOfMethodCallIgnored
-                tables.get(tableIndex); // will index-out-of-bounds if not valid
-                //noinspection ResultOfMethodCallIgnored
-                ranges.get(rangeIndex); // will index-out-of-bounds if not valid
-
-                SortedArrayList<Node.Id> nodes = CollectionSerializers.deserializeSortedArrayList(in, nodeId, Node.Id[]::new);
-                BitSet notInFastPath = BitSetSerializer.instance.deserialize(in);
-                BitSet joining = BitSetSerializer.instance.deserialize(in);
-                int flags = in.readUnsignedVInt32();
-                shards.add(new CompactTopology.ShardRef(tableIndex, rangeIndex, nodes, notInFastPath, joining, new TinyEnumSet<>(flags)));
-            }
-            return new CompactTopology(epoch, staleNodes, tables, ranges, shards);
-        }
-    };
 
     public static final UnversionedSerializer<Topology> topology = new UnversionedSerializer<>()
     {
